# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API untuk personal portfolio (projects, experiences, contact, dll).",
        "title": "Portfolio Backend API",
        "contact": {
            "name": "Fauzan",
            "email": "fauzan@pandi.id"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/projects": {
            "get": {
                "description": "List projects visible publicly with search \u0026 pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Get public projects",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter featured",
                        "name": "featured",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ProjectsListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/projects/{slug}": {
            "get": {
                "description": "Get single public project by slug",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Get project detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ProjectResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "details": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                },
                "requestId": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ProjectFeatureResponse": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        },
        "handlers.ProjectResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "challenge": {
                    "type": "string"
                },
                "coverImageUrl": {
                    "type": "string"
                },
                "demoUrl": {
                    "type": "string"
                },
                "features": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ProjectFeatureResponse"
                    }
                },
                "id": {
                    "type": "string"
                },
                "isFeatured": {
                    "type": "boolean"
                },
                "longDescription": {
                    "type": "string"
                },
                "repoUrl": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "role": {
                    "type": "string"
                },
                "screenshots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ProjectScreenshotResponse"
                    }
                },
                "shortDesc": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "solution": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.TagResponse"
                    }
                },
                "technicalDetails": {},
                "timeline": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "handlers.ProjectScreenshotResponse": {
            "type": "object",
            "properties": {
                "imageUrl": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.ProjectsListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ProjectResponse"
                    }
                },
                "meta": {}
            }
        },
...
        "handlers.TagResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}